631: (739?) Fix JSON output when -class is used and/or other weirdness.
This fix refactors all the if(func()=="")... duplicates to a series of
polymorphic JSON snippet generators.  It also sets up a small state tracker
for whether a comma should be prepended to the next entry (which fixes a bunch
of oddball cases where commas are missing or extraneous that was breaking
JSON format).  Since this code was adapted from asXML (which has no need to
track commas), it's likely that the bug was vestigal from assumptions made in
asXML().

Hopefully this both fixes all the JSON format issues and makes the code much
easier to maintain going forward.